Move specified and computed time values to submodules
